+++ SprPlotM.c
@@ -22,6 +22,9 @@
   CJB: 07-May-25: Dogfooding the _Optional qualifier.
 */
 
+/* ISO library headers */
+#include <stdint.h>
+
 /* Acorn C/C++ library headers */
 #include "kernel.h"
 #include "swis.h"
@@ -57,13 +60,13 @@ _Optional _kernel_oserror *os_sprite_op_plot_trans_matrix_sprite(
 
   _kernel_swi_regs regs = {.r = {
      SPRITEOP_USERAREA_SPRNAME + SPRITEOP_PLOT_TRANSFORMED_SPRITE,
-     (int)area,
-     (int)name,
+     (intptr_t)area,
+     (intptr_t)name,
      flags,
-     src ? (int)src : 0,
+     src ? (intptr_t)src : 0,
      action,
-     (int)matrix,
-     colours ? (int)colours : 0
+     (intptr_t)matrix,
+     colours ? (intptr_t)colours : 0
   }};
 
   return _kernel_swi(OS_SpriteOp, &regs, &regs);
